backend/requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyJWT==2.10.1
bcrypt==4.2.1
pydantic==2.10.3
python-dotenv==1.0.1
pyodbc==5.2.0
SQLAlchemy==2.0.36
alembic==1.14.0
marshmallow==3.23.2
APScheduler==3.11.0

